feat(core): implements Render err/panic semantics
The Render method does not and should not return errors
by default. This commit implements semantics which guarantee
panics occur on server side bugs or errors, which are not
io.Writer errors. For example if Content-Length http headers
are set to 0, while content is actively written, a panic
will be thrown to notify the implementation (and avoid silently
discarding this)

However, io.Writer errors such as client disconnecting is silently
ignored as it is outside the servers responsibility, but can also
be captured by having the calling implementation wrap io.Writer
in a wrapper which captures the error. This has also been documented.