feat(livereload): simplifies setup and config
livereload now handles the close of the fsnotify watcher internally
by listening for the os.Interrupt signal, the interrupt signal
is also forwarded on to the main process to make it's dependency
transparent to the main process.

livereload also does not require a custom reload function
to be provided, if nil is provided a default handler is used instead.