fix(core): SetConfig propagates properly
SetConfig was not propagating new setting
adequately. This results in a config to be reset
but if the Templates have already been defined
on the previous settings, they would not propage
as one would expect.

The error was with the assignment to a pointer
instead of assignment to a value.